Make the date persistant between tabs
Resolves issue #4

I thought it'd be better to keep the URI style, and just change the functionality of the top links to use the start/end date in their urls. This achieves the required functionality of the request without changing the structure of the website, and keeping it query-string free.

I've also enabled the default 'home' functionality on the navbar-brand, which takes the user back to "/", which is the default entry point. Small css change because bootstrap thinks that a link in a navbar-brand should be glaringly ugly.