Remove form wrapper from navbar form
This is not a 'real' form. the '#reload.click` function from pigeonhole.js takes the form elements and contructs a new GET-based URL to navigate to. By wrapping in a form, a POST event can occur when enter is pressed on a text field. This does not allow for the capturing of the information in these fields server side.

So, if we remove the form, and create a js-based capture on enter on the search field (last commit), then we get the same functionality without the issue of different routing pathways