Super squishy navbar, batman!
When using a wide container (pre-defined as min-width 1200px), hardcode smaller field sizes so that, with the current navbar link names, the search elements and the navbar links all fit in one row.

Some.. eccentricities seen to exist between firefox and chrome in regards to a series of UI elements appearing chrome on mouseover that do not appear in firefox. If these are able to be removed or altered, then the datefields could be shortened again to allow for only the date and not these elements. However, this might be getting a little bit on the nit-picky side when we could just readdress if we need to have 'alert' appearing in every navbar link element.