Handle inserts on an empty database
`entries` on line 29 is assumed to return something, and therefore have a timeseries key to pass down.
However, on empty databases, there are no results, so `entries` on line 30 is `nil`, which breaks when `.select` is called later.

Setting `entries` as an empty array solves this issue, as the select doesn't fail as it does on `nil`

Before change: an empty database cannot be populated with the `./bin/import_from_pd` import script
After change: an empty database can be populated via the import script