Encoding never bothered me anyway
The encoding was originally to attempt to circumvent issues with malformed searches. Now that the ro/rw is in effect, this is not required. A search term with sufficient requirements (e.g. a space in the term) will be automatically encoded in the URI, but will search the database as is.